Added MVC structure for views and controllers
